Move DefaultServiceIPRange into pkg/master
